hacky way of hopefully fixing errors
